- New about dialog and icon :D - Updated readme - Added link to guide in menu - Cleaned up a little

+++ hkb_editor/gui/widgets/graph_widget.py
@@ -417,6 +417,9 @@ class GraphWidget:
             self.set_origin(0.0, 0.0)
 
     def regenerate(self):
+        if not self.graph:
+            return
+        
         dpg.delete_item(f"{self.tag}_edge_layer", children_only=True)
         dpg.delete_item(f"{self.tag}_node_layer", children_only=True)
         self.color_generator.reset()
